refactor: Replace user-service with users
