+++ users/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/users/Users.java
@@ -1,13 +1,13 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.userservice;
+package com.github.itsAkshayDubey.eventdrivenarchitecture.users;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class UserServiceApplication {
+public class Users {
 
 	public static void main(String[] args) {
-		SpringApplication.run(UserServiceApplication.class, args);
+		SpringApplication.run(Users.class, args);
 	}
 
 }
