+++ users/src/main/resources/application.properties
@@ -1,3 +1,3 @@
 server.port=8084
-spring.application.name=user-service
+spring.application.name=users
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka
